+++ lib/handlers/wk_item.py
@@ -3,63 +3,61 @@ import requests
 import time
 import json
 
-
-# Local storage definitions
-ITEM_BASE_DIR = "data/items"
-ITEM_BASE_NAME = "{item_type}.json"
-ITEM_CACHE_DIR = "data/items/{item_type}"
-ITEM_CACHE_NAME = "{item_type}_lvl{level}.json"
-
-# Header requests and endpoints
-HTTP_HEADERS = {
-    "Wanikani-Revision": "20170710",
-    "Authorization": "Bearer {wk_token}",
-}
-API_MAIN = "https://api.wanikani.com/v2/"
-API_SUBJECTS = "subjects"
+from lib.utils.constants import ITEM_TYPE, ITEM_DB_FILE
 
 
-def cache_items(wk_token: str, status: bool = False):
-    """Main function to cache WaniKani subject items."""
-    if status:
-        os.makedirs(ITEM_BASE_DIR, exist_ok=True)
+def find_item(char: str, database: list):
+    """Find an item from WaniKani items data."""
+    for item in database:
+        meanings = [meaning.lower() for meaning in item["meanings"]]
+        if char == item["char"] or char.lower() in meanings:
+            return item
+    return None
 
-        for item_type in ["radical", "kanji", "vocabulary"]:
-            print(f"Caching WaniKani {item_type} items:")
+
+def load_item_db(wk_token: str, item_dbs: dict, cache: bool = False):
+    """Main function to load WaniKani subject items to the bot."""
+    os.makedirs("data/items", exist_ok=True)
+    
+    if cache:
+        for item_type, name in ITEM_TYPE.items():
+            name = name.lower()
+            print(f"Caching WaniKani {name} items:")
 
             # Setup cache and base files and direcory
-            base_fp = os.path.join(ITEM_BASE_DIR, ITEM_BASE_NAME.format(item_type=item_type))
+            base_fp = ITEM_DB_FILE[item_type]
             file_status(base_fp)
-            cache_dir = ITEM_CACHE_DIR.format(item_type=item_type)
+            cache_dir = f"data/items/{name}"
             os.makedirs(cache_dir, exist_ok=True)
             
             # Cache items by level
             for level in range(1, 61):
-                cache_fn = ITEM_CACHE_NAME.format(item_type=item_type, level=level)
-                cache_fp = os.path.join(cache_dir, cache_fn)
+                cache_fp = os.path.join(cache_dir, f"{name}_lvl{level}.json")
                 file_status(cache_fp)
-
-                # Get and save cache file
-                cache_data: list = get_cache_item(wk_token, item_type, level)
-                with open(cache_fp, 'wb') as file:
-                    file.write(json.dumps(cache_data, ensure_ascii=False, indent=4).encode("utf8"))
+                cache_data = cache_item(wk_token, name, level)
+                with open(cache_fp, 'wb') as output:
+                    output.write(json.dumps(cache_data, ensure_ascii=False, indent=4, default=str).encode("utf8"))
 
             # Combine cache files
-            base_data = []
-            cache_files = os.listdir(cache_dir)
-            for cache_file in cache_files:
-                cache_fp = os.path.join(cache_dir, cache_file)
-                with open(cache_fp, 'r') as input:
-                    base_data.extend(json.load(input))
+            item_db = []
+            for cache_file in os.listdir(cache_dir):
+                with open(os.path.join(cache_dir, cache_file), 'r') as input:
+                    item_db.extend(json.load(input))
             with open(base_fp, 'wb') as output:
-                output.write(json.dumps(base_data, ensure_ascii=False, indent=4).encode("utf8"))
+                output.write(json.dumps(item_db, ensure_ascii=False, indent=4, default=str).encode("utf8"))
             
-            print(f"Finished caching {item_type} items.\n")
+            print(f"Finished caching {name} items.")
     else:
-        print("WaniKani items update not initialized.")
+        print("Items not renewed.")
+    
+    # Now load the data
+    for type, file in ITEM_DB_FILE.items():
+        with open(file, "r") as data:
+            item_dbs[type] = json.load(data)
+            print(f"- {file} loaded")
 
 
-def file_status(filepath: str) -> str:
+def file_status(filepath: str):
     """Indicator whether the file is new or existing."""
     if os.path.isfile(filepath):
         print(f"! Overwriting '{filepath}'")
@@ -67,13 +65,14 @@ def file_status(filepath: str) -> str:
         print(f"- Creating '{filepath}'")
 
 
-def filter_cache_item(data: list, item_type: str) -> dict:
+def filter_raw_cache(data: list, item_type: str) -> dict:
     """Filter raw WaniKani item data for the given item type."""
     if item_type == "radical":
-        char_image = None
         for image in data["data"]["character_images"]:
             if image["content_type"] == "image/png" and image["metadata"]["style_name"] == "original":
                 char_image = image["url"]
+            else:
+                char_image = None
         meanings = [meaning["meaning"] for meaning in data["data"]["meanings"]]
         return {
             "id": data["id"],
@@ -129,30 +128,16 @@ def filter_cache_item(data: list, item_type: str) -> dict:
         }
 
 
-def get_cache_item(wk_token: str, item_type: str, level: int) -> list:
+def cache_item(wk_token: str, item_type: str, level: int) -> list:
     """Requests WaniKani subjects using its API based on item_type and level."""
-
-    # Load API token into headers
-    HTTP_HEADERS["Authorization"] = HTTP_HEADERS["Authorization"].format(
-        wk_token=wk_token)
-
-    # Assemble query
-    url = API_MAIN + API_SUBJECTS
-    if item_type:
-        if level:
-            url += f"?types={item_type}&levels={level}"
-        else:
-            url += f"?types={item_type}"
-    elif level:
-        url += f"?levels={level}"
-
     # Execute requests
-    req = requests.get(url, headers=HTTP_HEADERS)
+    endpoint = f"https://api.wanikani.com/v2/subjects?types={item_type}&levels={level}"
+    req = requests.get(endpoint, headers={
+        "Authorization": f"Bearer {wk_token}"
+    })
     time.sleep(0.5)
 
     # Filter output data based on item_type
-    data: list = req.json()["data"]
-    data_filtered: list = [filter_cache_item(
-        field, item_type) for field in data]
-
+    data_raw = req.json()["data"]
+    data_filtered = [filter_raw_cache(item, item_type) for item in data_raw]
     return data_filtered