+++ lib/utils/constants.py
@@ -1,0 +1,36 @@
+ITEM_TYPE = {
+    'rad': 'Radical',
+    'kan': 'Kanji',
+    'voc': 'Vocabulary'
+}
+
+# to be used in bot.item_data
+ITEM_DB_FILE = {
+    'rad': "data/items/radical.json",
+    'kan': "data/items/kanji.json",
+    'voc': "data/items/vocabulary.json"
+}
+
+MNEMONIC_TYPE = {
+    'mean': 'Meaning',
+    'read': 'Reading',
+    'both': 'Both'
+}
+
+LICENSE_TYPE = {
+    'midj_fr': 'MidJourney (free)',
+    'midj_pd': 'MidJourney (paid)',
+    'dalle_2': 'DALL-E 2',
+    'ot_comm': 'Other AI (commercially usable)',
+    'ot_prsn': 'Other AI (personal use)',
+    'drawing': 'Own drawing',
+}
+
+EMBED_COLOR = {
+    'pink': 0xe377c2,
+    'red': 0xd62728,
+    'green': 0x2ca02c,
+    'rad': 0x00AAFF,
+    'kan': 0xFF00AA,
+    'voc': 0xAA00FF
+}