+++ main.py
@@ -9,11 +9,10 @@ from discord import Intents, Bot, Embed
 from discord.ext import tasks, commands
 
 import lib.exceptions as exceptions
-from lib.utils.manage_item import ITEM_NAMES, load_item_data
-from lib.utils.scrape_wk import cache_items
+from lib.handlers.wk_item import load_item_db, load_item_db
 
 
-VERSION = '0.2.2'
+VERSION = '0.2.3'
 UPDATE_WK_ITEMS = False
 
 
@@ -39,12 +38,12 @@ bot.version = VERSION
 @bot.event
 async def on_ready() -> None:
     """Called when the connection to Discord has been established."""
-    print("-------------------")
+    print("\n-------------------")
     print(f"Logged in as '{bot.user}' v{bot.version}")
     print(f"PyCord API version: {discord.__version__}")
     print(f"Python version: {platform.python_version()}")
     print(f"Running on: {platform.system()} {platform.release()} ({os.name})")
-    print("-------------------")
+    print("-------------------\n")
     bot_presence.start()
 
     embed = Embed(
@@ -74,32 +73,9 @@ async def bot_presence() -> None:
     await bot.change_presence(activity=discord.Game(random.choice(statuses)))
 
 
-if __name__ == "__main__":
-    # Loading bot cogs
-    print("\nLoading Crabigator cogs:")
-    for file in os.listdir(f"./lib/cogs"):
-        if file.endswith(".py"):
-            cog = file[:-3]
-            try:
-                bot.load_extension(f"lib.cogs.{cog}")
-                print(f"- {cog} cog loaded")
-            except Exception as e:
-                exception = f"{type(e).__name__}: {e}"
-                print(f"! {cog} cog failed to load:\n  {exception}")
-    print("Finished loading the cogs.\n")
-
-    # Loading WaniKani items
-    cache_items(config["wk_token"], UPDATE_WK_ITEMS)
-    bot.item_names = ITEM_NAMES
-    bot.item_data = {}
-    print("\nLoading WaniKani items:")
-    load_item_data(bot)
-    print("Finished loading all items.\n")
-
-
 @bot.event
 async def on_application_command(ctx) -> None:
-    """Executed every time a slash command has been *successfully* executed."""
+    """Executed every time a slash command has been successfully executed."""
     message = (
         f"Executed '/{ctx.command.name}' command in "
         f"{ctx.guild.name} (ID: {ctx.guild.id}) by "
@@ -110,10 +86,7 @@ async def on_application_command(ctx) -> None:
 
 @bot.event
 async def on_application_command_error(ctx, exception) -> None:
-    """Executed every time a valid slash command catches an error.
-
-    'ephemeral=True' will make so that only the user who execute the command can see the message.
-    """
+    """Executed every time a valid slash command catches an error."""
     if isinstance(exception, commands.CommandOnCooldown):
         minutes, seconds = divmod(exception.retry_after, 60)
         hours, minutes = divmod(minutes, 60)
@@ -124,28 +97,28 @@ async def on_application_command_error(ctx, exception) -> None:
             f"{f'{round(minutes)} minutes' if round(minutes) > 0 else ''} "
             f"{f'{round(seconds)} seconds' if round(seconds) > 0 else ''}."
         )
-        embed = discord.Embed(
+        embed = Embed(
             title="Hey, please slow down!",
             description=message,
             color=0xFF0000
         )
         return await ctx.send(embed=embed, ephemeral=True)
     elif isinstance(exception, exceptions.UserBlacklisted):
-        embed = discord.Embed(
+        embed = Embed(
             title="Error!",
             description="You are blacklisted from using the bot.",
             color=0xFF0000
         )
         return await ctx.send(embed=embed, ephemeral=True)
     elif isinstance(exception, exceptions.UserNotDeveloper):
-        embed = discord.Embed(
+        embed = Embed(
             title="Error!",
             description="You are not the developer of the bot!",
             color=0xFF0000
         )
         return await ctx.send(embed=embed, ephemeral=True)
     elif isinstance(exception, commands.MissingPermissions):
-        embed = discord.Embed(
+        embed = Embed(
             title="Error!",
             description="You are missing the permission(s) `" + ", ".join(
                 exception.missing_permissions) + "` to execute this command!",
@@ -153,7 +126,7 @@ async def on_application_command_error(ctx, exception) -> None:
         )
         return await ctx.send(embed=embed, ephemeral=True)
     elif isinstance(exception, commands.MissingRequiredArgument):
-        embed = discord.Embed(
+        embed = Embed(
             title="Error!",
             # Command arguments have no capital letter in the code.
             description=str(exception).capitalize(),
@@ -161,6 +134,27 @@ async def on_application_command_error(ctx, exception) -> None:
         )
         await ctx.send(embed=embed, ephemeral=True)
     raise exception
+
+
+if __name__ == "__main__":
+    # Loading bot cogs
+    print("\nLoading Crabigator cogs:")
+    for file in os.listdir("lib/cogs"):
+        if file.endswith(".py"):
+            cog = file[:-3]
+            try:
+                bot.load_extension(f"lib.cogs.{cog}")
+                print(f"- {cog} cog loaded")
+            except Exception as e:
+                exception = f"{type(e).__name__}: {e}"
+                print(f"! {cog} cog failed to load:\n  {exception}")
+    print("Finished loading the cogs.")
+
+    # Loading WaniKani items
+    print("\nLoading WaniKani items:")
+    bot.item_data = {}
+    load_item_db(config["wk_token"], bot.item_data, UPDATE_WK_ITEMS)
+    print("Finished loading all items.")
 
 
 # Finally run the bot